Fix data type conversion when instantiating a Matrix
When initialising a `Matrix` object, the default `dtype` is currently `int`.
This means that when the `dtype` was not specified unexpected things could happen such as:
```py
> from discopy.matrix import Matrix
> Matrix([0.5, 0.5], dom=1, cod=2)
Matrix([0, 0], dom=1, cod=2)
```

This PR addresses this issue by using the `dtype` of the created `np.array` when `dtype` is not specified for the `Matrix`.

Also fixing a Warning in a test.