Change the way the lxml is imported
The package lxml is only required for a specific drawing function, Grid.to_html in grid.py.
Because of this, lxml is not listed in requirements.txt, and it has been imported inside a try-except block.
However, this results in very poor error messages:
```
NameError: name 'Element' is not defined
```

This commit places the lxml import inside the function where it is required so that we have a better error message:
```
ModuleNotFoundError: No module named 'lxml'
```
Also, a not has been added to the docstring of the function where it is imported.