Update the dependencies of the documentation
The dependencies have been updated as `m2r2` has a dependency that just wouldn't work with a moder setup.

In newer versions of `sphinx`, it is possible to simply include markdown files given that `myst-parser` is included as an extension.

Some warnings of sphynx have been resolved, but there's still a lot to clean up.

Broken links have been fixed.